sicora-data-loader/requirements.txt
streamlit==1.29.0
pandas==2.1.3
psycopg2-binary==2.9.9
openpyxl==3.1.2
xlrd==2.0.1
tabula-py==2.9.0
PyPDF2==3.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
loguru==0.7.2
